Moved code from session_items to trello_cards
