+++ app.py
@@ -1,6 +1,5 @@
 from flask import Flask, redirect, render_template, request, url_for
 import trello_cards
-import session_items as session
 import os
 from card import Card
 
@@ -8,7 +7,7 @@ app = Flask(__name__)
 
 @app.route('/', methods=['GET'])
 def get_items():
-    trello_todos = session.get_items()
+    trello_todos = trello_cards.get_items()
 
     cards = []
     for trello_todo in trello_todos:
@@ -28,25 +27,25 @@ def add_item():
     name = request.form['name']
     desc = request.form['desc']
 
-    session.add_new_item(name, desc)
+    trello_cards.add_new_item(name, desc)
     return redirect('/')
 
 @app.route('/done', methods=['GET'])
 def get_done_items():
-    done = session.get_done_items()
+    done = trello_cards.get_done_items()
 
     return render_template('done.html', dones=done)
 
 @app.route('/done/<id>', methods=['POST'])
 def update_item(id):
-    session.update_item(id)
+    trello_cards.update_item(id)
 
     return redirect('/done')
 
 
 @app.route('/delete_item/<id>', methods=['POST'])
 def delete_item(id):
-    session.delete_item(id)
+    trello_cards.delete_item(id)
 
     # return render_template('index.html')
     return redirect("/")
